+++ src/main/java/com/booking/repository/TaxRepository.java
@@ -4,7 +4,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.booking.bean.Tax;
 
-public interface TaxRepository extends JpaRepository<Tax, Integer>{
+public interface TaxRepository extends JpaRepository<Tax, Long>{
 
 	Tax findAllByTaxid(Long taxid);
 }
